Implement v2 APIs (#22)
In short: this PR drops the redundant "motion" prefix on various types and aligns naming with system terminology where possible and re-implements the APIs as Objective-C APIs, closing #5.

| Old API | New API | Rationale |
|:------- |:-------- |:-----------|
| MotionTiming | AnimationTraits | This structure is intended to describe animations only, not motion in general. |
| MotionCurve | TimingCurve | This brings the API name closer to the similarly-purposed `CAMediaTimingFunction`. MotionCurve could also be easily confused with motion through x/y space rather than through time (e.g. ArcMove), which will be problematic as we start defining paths of motion through space. |
| MotionRepetition | RepetitionTraits | This aligns the naming with AnimationTraits. |

All prior APIs have been deleted, including the macro-based APIs and any active deprecations. This is a major change and will bump the release to v2.

A migration script has been included to ease migration of existing code.